Split out deduceImageWidthAndHeight
